+++ mock/attesterduties.go
@@ -21,8 +21,17 @@ import (
 )
 
 // AttesterDuties obtains attester duties.
-// If validatorIndicess is nil it will return all duties for the given epoch.
-func (*Service) AttesterDuties(_ context.Context, opts *api.AttesterDutiesOpts) (*api.Response[[]*apiv1.AttesterDuty], error) {
+// If validatorIndices is nil it will return all duties for the given epoch.
+func (s *Service) AttesterDuties(ctx context.Context,
+	opts *api.AttesterDutiesOpts,
+) (
+	*api.Response[[]*apiv1.AttesterDuty],
+	error,
+) {
+	if s.AttesterDutiesFunc != nil {
+		return s.AttesterDutiesFunc(ctx, opts)
+	}
+
 	data := make([]*apiv1.AttesterDuty, len(opts.Indices))
 	for i := range opts.Indices {
 		data[i] = &apiv1.AttesterDuty{