+++ mock/proposerduties.go
@@ -22,7 +22,11 @@ import (
 
 // ProposerDuties obtains proposer duties for the given epoch.
 // If validatorIndices is empty all duties are returned, otherwise only matching duties are returned.
-func (*Service) ProposerDuties(_ context.Context, opts *api.ProposerDutiesOpts) (*api.Response[[]*apiv1.ProposerDuty], error) {
+func (s *Service) ProposerDuties(ctx context.Context, opts *api.ProposerDutiesOpts) (*api.Response[[]*apiv1.ProposerDuty], error) {
+	if s.ProposerDutiesFunc != nil {
+		return s.ProposerDutiesFunc(ctx, opts)
+	}
+
 	data := make([]*apiv1.ProposerDuty, len(opts.Indices))
 	for i := range opts.Indices {
 		data[i] = &apiv1.ProposerDuty{