+++ mock/validators.go
@@ -22,12 +22,16 @@ import (
 )
 
 // Validators provides the validators, with their balance and status, for a given state.
-func (*Service) Validators(_ context.Context,
-	_ *api.ValidatorsOpts,
+func (s *Service) Validators(ctx context.Context,
+	opts *api.ValidatorsOpts,
 ) (
 	*api.Response[map[phase0.ValidatorIndex]*apiv1.Validator],
 	error,
 ) {
+	if s.ValidatorsFunc != nil {
+		return s.ValidatorsFunc(ctx, opts)
+	}
+
 	return &api.Response[map[phase0.ValidatorIndex]*apiv1.Validator]{
 		Data:     map[phase0.ValidatorIndex]*apiv1.Validator{},
 		Metadata: make(map[string]any),