+++ testclients/sleepy.go
@@ -181,11 +181,11 @@ func (s *Sleepy) TargetAggregatorsPerCommittee(ctx context.Context) (uint64, err
 	return next.TargetAggregatorsPerCommittee(ctx)
 }
 
-// AggregateAttestation fetches the aggregate attestation given an attestation.
+// AggregateAttestation fetches the aggregate attestation for the given options.
 func (s *Sleepy) AggregateAttestation(ctx context.Context,
 	opts *api.AggregateAttestationOpts,
 ) (
-	*api.Response[*phase0.Attestation],
+	*api.Response[*spec.VersionedAttestation],
 	error,
 ) {
 	s.sleep(ctx)
@@ -198,14 +198,14 @@ func (s *Sleepy) AggregateAttestation(ctx context.Context,
 }
 
 // SubmitAggregateAttestations submits aggregate attestations.
-func (s *Sleepy) SubmitAggregateAttestations(ctx context.Context, aggregateAndProofs []*phase0.SignedAggregateAndProof) error {
+func (s *Sleepy) SubmitAggregateAttestations(ctx context.Context, opts *api.SubmitAggregateAttestationsOpts) error {
 	s.sleep(ctx)
 	next, isNext := s.next.(consensusclient.AggregateAttestationsSubmitter)
 	if !isNext {
 		return errors.New("next does not support this call")
 	}
 
-	return next.SubmitAggregateAttestations(ctx, aggregateAndProofs)
+	return next.SubmitAggregateAttestations(ctx, opts)
 }
 
 // AttestationData fetches the attestation data for the given slot and committee index.
@@ -241,7 +241,7 @@ func (s *Sleepy) AttestationPool(ctx context.Context,
 }
 
 // SubmitAttestations submits attestations.
-func (s *Sleepy) SubmitAttestations(ctx context.Context, attestations []*phase0.Attestation) error {
+func (s *Sleepy) SubmitAttestations(ctx context.Context, attestations *api.SubmitAttestationsOpts) error {
 	s.sleep(ctx)
 	next, isNext := s.next.(consensusclient.AttestationsSubmitter)
 	if !isNext {
@@ -252,7 +252,7 @@ func (s *Sleepy) SubmitAttestations(ctx context.Context, attestations []*phase0.
 }
 
 // AttesterDuties obtains attester duties.
-// If validatorIndicess is nil it will return all duties for the given epoch.
+// If validatorIndices is nil it will return all duties for the given epoch.
 func (s *Sleepy) AttesterDuties(ctx context.Context,
 	opts *api.AttesterDutiesOpts,
 ) (
@@ -640,4 +640,52 @@ func (s *Sleepy) BlobSidecars(ctx context.Context,
 	}
 
 	return next.BlobSidecars(ctx, opts)
+}
+
+// AttestationRewards provides rewards to the given validators for attesting.
+func (s *Sleepy) AttestationRewards(ctx context.Context,
+	opts *api.AttestationRewardsOpts,
+) (
+	*api.Response[*apiv1.AttestationRewards],
+	error,
+) {
+	s.sleep(ctx)
+	next, isNext := s.next.(consensusclient.AttestationRewardsProvider)
+	if !isNext {
+		return nil, fmt.Errorf("%s@%s does not support this call", s.next.Name(), s.next.Address())
+	}
+
+	return next.AttestationRewards(ctx, opts)
+}
+
+// BlockRewards provides rewards for proposing a block.
+func (s *Sleepy) BlockRewards(ctx context.Context,
+	opts *api.BlockRewardsOpts,
+) (
+	*api.Response[*apiv1.BlockRewards],
+	error,
+) {
+	s.sleep(ctx)
+	next, isNext := s.next.(consensusclient.BlockRewardsProvider)
+	if !isNext {
+		return nil, fmt.Errorf("%s@%s does not support this call", s.next.Name(), s.next.Address())
+	}
+
+	return next.BlockRewards(ctx, opts)
+}
+
+// SyncCommitteeRewards provides rewards to the given validators for being members of a sync committee.
+func (s *Sleepy) SyncCommitteeRewards(ctx context.Context,
+	opts *api.SyncCommitteeRewardsOpts,
+) (
+	*api.Response[[]*apiv1.SyncCommitteeReward],
+	error,
+) {
+	s.sleep(ctx)
+	next, isNext := s.next.(consensusclient.SyncCommitteeRewardsProvider)
+	if !isNext {
+		return nil, fmt.Errorf("%s@%s does not support this call", s.next.Name(), s.next.Address())
+	}
+
+	return next.SyncCommitteeRewards(ctx, opts)
 }